backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
# Pin bcrypt to avoid passlib compatibility issues on Windows
bcrypt==4.1.3
python-multipart==0.0.12
requests==2.32.3
# For tools/e2e scripts only; keep >=13 to avoid conflicts with yfinance
websockets==13.0.1
aiofiles==24.1.0

# === 第二阶段新增依赖 ===
# PostgreSQL驱动
psycopg2-binary==2.9.9
# Redis异步客户端
redis==5.0.1
# HTTP异步客户端(用于告警通知)
httpx==0.27.0
# 数据库迁移工具
alembic==1.13.1

